requirements.txt
# Workspace Dependencies

# Core
pyyaml>=6.0
requests>=2.28.0

# Dashboard
streamlit>=1.28.0

# Optional - for future features
# fastapi>=0.100.0      # For API interface
# uvicorn>=0.23.0       # ASGI server
# python-telegram-bot   # For Telegram interface
# rich>=13.0.0          # For pretty CLI output
